Extracted methods and implemented as page-rendering on habits.js page
